ref(metrics): Split aggregator into multiple modules
